Increased version, updated unit tests and removed default unit test.
Increasing the spec file version due to added features. Since the messages are sent more often now the unit test required updating.
Ive removed the default unit test since it tries to test the arm implementation and in general that will not be present unless
compiling for the arm platform.